+++ GeometryNative/OdGePoint2d.h
@@ -7,10 +7,10 @@ namespace GeometryNative
 	public:
 		static const OdGePoint2d kOrigin;
 		OdGePoint2d();
-		OdGePoint2d(float x, float y);
+		OdGePoint2d(double x, double y);
 		~OdGePoint2d();
 		double DistanceTo(OdGePoint2d point) const;
-		float x;
-		float y;
+		double x;
+		double y;
 	};
 }
