+++ GeometryNative/OdGeVector2d.cpp
@@ -4,21 +4,21 @@
 namespace GeometryNative
 {
     OdGeVector2d::OdGeVector2d() : x(0), y(0) {}
-    OdGeVector2d::OdGeVector2d(float x, float y) : x(x), y(y) {}
+    OdGeVector2d::OdGeVector2d(double x, double y) : x(x), y(y) {}
     OdGeVector2d::~OdGeVector2d() {}
 
-    float OdGeVector2d::Length() const
+    double OdGeVector2d::Length() const
     {
         return sqrt(x * x + y * y);
     }
 
     OdGeVector2d OdGeVector2d::Normalize() const
     {
-        float length = Length();
+        double length = Length();
         return OdGeVector2d(x / length, y / length);
     }
 
-    float OdGeVector2d::DotProduct(const OdGeVector2d& other) const
+    double OdGeVector2d::DotProduct(const OdGeVector2d& other) const
     {
         return x * other.x + y * other.y;
     }
@@ -45,7 +45,7 @@ namespace GeometryNative
         return OdGeVector2d(x - other.x, y - other.y);
     }
 
-    OdGeVector2d OdGeVector2d::operator*(float scalar) const
+    OdGeVector2d OdGeVector2d::operator*(double scalar) const
     {
         return OdGeVector2d(x * scalar, y * scalar);
     }
