+++ MathGL/MathGL.cpp
@@ -47,8 +47,8 @@ void pickObject(int x, int y);
 void setColorID(int id);
 
 int win_width, win_height;
-float cam_theta, cam_phi = 25, cam_dist = 8;
-float cam_pan[3];
+double cam_theta, cam_phi = 25, cam_dist = 8;
+double cam_pan[3];
 int mouse_x, mouse_y;
 int bnstate[8];
 int anim, help;
@@ -126,7 +126,7 @@ HWND MathGL::GLEngineNative::InitializeWindow(HINSTANCE hInstance, int nCmdShow,
 	return hwnd;
 }
 
-void drawGridXY(bool picking = false, float size = 10.0f, float step = 1.0f)
+void drawGridXY(bool picking = false, double size = 10.0f, double step = 1.0f)
 {
 	if (picking)
 	{
@@ -143,7 +143,7 @@ void drawGridXY(bool picking = false, float size = 10.0f, float step = 1.0f)
 		glColor4f(0.5f, 0.5f, 0.5f, 0.5f);
 
 	glColor4f(0.5f, 0.5f, 0.5f, 0.5f);
-	for (float i = step; i <= size; i += step)
+	for (double i = step; i <= size; i += step)
 	{
 		glVertex3f(-size, i, 0);
 		glVertex3f(size, i, 0);
@@ -173,7 +173,7 @@ void drawGridXY(bool picking = false, float size = 10.0f, float step = 1.0f)
 }
 
 
-void drawAxis(bool picking = false, float size = 2.5f)
+void drawAxis(bool picking = false, double size = 2.5f)
 {
 	glLineWidth(2.0f);
 
@@ -233,7 +233,7 @@ void display(void)
 void drawScene(bool picking)
 {
 	long tm;
-	float lpos[] = { -1, 2, 3, 0 };
+	double lpos[] = { -1, 2, 3, 0 };
 
 	glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
 	glMatrixMode(GL_MODELVIEW);
@@ -385,7 +385,7 @@ void print_help(void)
 #define ZNEAR	0.5f
 void reshape(int x, int y)
 {
-	float vsz, aspect = (float)x / (float)y;
+	double vsz, aspect = (double)x / (double)y;
 	win_width = x;
 	win_height = y;
 
@@ -485,9 +485,9 @@ void motion(int x, int y)
 		glutPostRedisplay();
 	}
 	if (bnstate[1]) {
-		float up[3], right[3];
-		float theta = cam_theta * M_PI / 180.0f;
-		float phi = cam_phi * M_PI / 180.0f;
+		double up[3], right[3];
+		double theta = cam_theta * M_PI / 180.0f;
+		double phi = cam_phi * M_PI / 180.0f;
 
 		up[0] = -sin(theta) * sin(phi);
 		up[1] = -cos(phi);
